Meal plans are now properly deleted

+++ ovs/services/resident_service.py
@@ -6,6 +6,7 @@ import logging
 from sqlalchemy.exc import SQLAlchemyError
 
 from ovs import db
+from ovs.services.meal_service import MealService
 from ovs.models.profile_model import Profile
 from ovs.models.resident_model import Resident
 from ovs.models.user_model import User
@@ -92,7 +93,10 @@ class ResidentService:
         from ovs.services.profile_service import ProfileService
         resident = ResidentService.get_resident_by_id(user_id)
         if resident is not None:
-            if ProfileService.delete_profile(user_id):
+            meal_delete = True
+            if resident.mealplan_pin != 0:
+                meal_delete = MealService.delete_meal_plan(resident.mealplan_pin)
+            if ProfileService.delete_profile(user_id) and meal_delete:
                 try:
                     db.session.delete(resident)
                     return True
